requirements.txt
HyperPyYAML==1.1.0
PyYAML==6.0
boto3==1.26.36
botocore==1.29.36
certifi==2022.12.7
filelock==3.8.2
huggingface-hub==0.11.1
idna==3.4
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
packaging==22.0
pip==21.1.1
psycopg2-binary==2.9.5
pyaudioconvert==0.0.5
pydub==0.25.1
python-dateutil==2.8.2
regex==2022.10.31
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sentencepiece==0.1.97
setuptools==56.0.0
six==1.16.0
speechbrain==0.5.13
tokenizers==0.13.2
torchaudio==0.13.1
tqdm==4.64.1
transformers==4.25.1
urllib3==1.26.13
